Feature/add security

+++ src/main/java/com/dotoku/carhop/service/UserService.java
@@ -1,11 +1,14 @@
 package com.dotoku.carhop.service;
 
-import com.dotoku.carhop.dto.HopUserDto;
-import com.dotoku.carhop.dto.VehicleDto;
+import com.dotoku.carhop.dto.*;
 import com.dotoku.carhop.dto.mapper.UserMapper;
+import com.dotoku.carhop.entity.HopRequest;
 import com.dotoku.carhop.entity.HopUser;
+import com.dotoku.carhop.entity.Review;
 import com.dotoku.carhop.entity.Vehicle;
 import com.dotoku.carhop.repository.UserRepository;
+import com.dotoku.carhop.security.JwtUtil;
+import com.dotoku.carhop.security.SecurityUtil;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +17,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,15 +30,17 @@ import java.util.stream.Collectors;
 public class UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
-
-
+    private final PasswordEncoder passwordEncoder;
+    private final JwtUtil jwtUtil;
 
     @Transactional
-    public ResponseEntity<HopUserDto> addUser(HopUserDto userDto){
+    public ResponseEntity<HopUserDto> addUser(HopUserRequestDto userDto){
         HopUser hopUser = userMapper.mapDtoToEntity(userDto);
+        hopUser.setPassword(passwordEncoder.encode(userDto.getPassword()));
         userRepository.save(hopUser);
         userDto.setId(hopUser.getId());
-        return ResponseEntity.ok(userDto);
+        HopUserDto hopUserDto = userMapper.mapEntityToDto(hopUser);
+        return ResponseEntity.ok(hopUserDto);
     }
 
     public ResponseEntity<HopUserDto> getUser(Long userId){
@@ -48,6 +54,11 @@ public class UserService {
         HopUser hopUser = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found!"));
 
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopUser.getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to update.");
+        }
+
         userMapper.mapDtoToEntity(userDto, hopUser);
         userRepository.save(hopUser);
         userDto.setId(hopUser.getId());
@@ -57,6 +68,11 @@ public class UserService {
     public ResponseEntity<String> deleteUser(Long userId) {
         HopUser hopUser = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found!"));
+
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopUser.getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to delete this account.");
+        }
 
         userRepository.delete(hopUser);
         return ResponseEntity.ok("User deleted successfully.");
@@ -74,6 +90,12 @@ public class UserService {
     public ResponseEntity<String> updateVehicle(VehicleDto vehicleDto, long userId) {
         HopUser hopUser = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found!"));
+
+        String loggedInEmail = SecurityUtil.getLoggedInEmail();
+        if (!hopUser.getEmail().equals(loggedInEmail)) {
+            throw new IllegalStateException("You are not authorized to update this vehicle.");
+        }
+
         Vehicle vehicle = hopUser.getVehicle();
 
         vehicle.setMake(vehicleDto.getMake());
@@ -87,6 +109,18 @@ public class UserService {
         return ResponseEntity.ok("Vehicle updated successfully.");
     }
 
+    public ResponseEntity<String> loginUser(String email, String password) {
+        HopUser user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException("User not found!"));
 
+        // Validate password
+        if (!passwordEncoder.matches(password, user.getPassword())) {
+            throw new RuntimeException("Invalid credentials");
+        }
+
+        // Generate JWT
+        String token = jwtUtil.generateToken(user);
+        return ResponseEntity.ok(token);
+    }
 
 }
